+++ src/test/java/com/techatpark/sjson/core/parser/NumberParserTest.java
@@ -1,6 +1,7 @@
-package com.techatpark.sjson.core;
+package com.techatpark.sjson.core.parser;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.techatpark.sjson.core.parser.NumberParser;
+import com.techatpark.sjson.core.Json;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
